refactor(persistence): Rehome persistence helpers
Move the persistence helper cluster from network.crypta.node to network.crypta.runtime.persistence in preparation for runtime-node extraction.

Make Persister public for cross-package use, add read-only file-path accessors, move the package-local tests with the helpers, update affected imports, add runtime persistence package documentation, and update the SpotBugs exclusion for ConfigurablePersisterParams.
