+++ src/main/java/network/crypta/node/NodeStatsConfig.java
@@ -8,6 +8,9 @@ import network.crypta.config.Option;
 import network.crypta.config.SubConfig;
 import network.crypta.l10n.NodeL10n;
 import network.crypta.runtime.bootstrap.NodeStarter;
+import network.crypta.runtime.persistence.ConfigurablePersister;
+import network.crypta.runtime.persistence.ConfigurablePersisterParams;
+import network.crypta.runtime.persistence.Persister;
 import network.crypta.support.SimpleFieldSet;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
