+++ src/main/java/network/crypta/runtime/endpoints/NodeClientPersistence.java
@@ -15,14 +15,14 @@ import network.crypta.client.async.persistence.PersistentRequestRecoveryCodec;
 import network.crypta.config.Option;
 import network.crypta.config.SubConfig;
 import network.crypta.crypt.MasterSecret;
-import network.crypta.node.ConfigurablePersister;
-import network.crypta.node.ConfigurablePersisterParams;
 import network.crypta.node.Node;
 import network.crypta.node.NodeClientCore;
 import network.crypta.node.NodeInitException;
-import network.crypta.node.Persistable;
 import network.crypta.runtime.endpoints.fcp.FcpEndpointHandle;
 import network.crypta.runtime.endpoints.fcp.FcpPersistentRequestServices;
+import network.crypta.runtime.persistence.ConfigurablePersister;
+import network.crypta.runtime.persistence.ConfigurablePersisterParams;
+import network.crypta.runtime.persistence.Persistable;
 import network.crypta.runtime.spi.RuntimePorts;
 import network.crypta.support.SimpleFieldSet;
 import network.crypta.support.Ticker;
