+++ src/test/java/network/crypta/runtime/persistence/ConfigurablePersisterTest.java
@@ -1,4 +1,4 @@
-package network.crypta.node;
+package network.crypta.runtime.persistence;
 
 import java.io.File;
 import java.io.IOException;
@@ -8,6 +8,7 @@ import network.crypta.config.InvalidConfigValueException;
 import network.crypta.config.Option;
 import network.crypta.config.SubConfig;
 import network.crypta.l10n.NodeL10n;
+import network.crypta.node.NodeInitException;
 import network.crypta.support.Ticker;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -56,10 +57,9 @@ class ConfigurablePersisterTest {
     File expectedTemp = new File(expectedTarget + ".tmp");
     assertTrue(expectedTarget.exists(), "target file should be created");
     assertTrue(expectedTemp.exists(), "temp file should be created");
-    assertEquals(expectedTarget.getCanonicalFile(), cp.persistTarget.getCanonicalFile());
-    assertEquals(expectedTemp.getCanonicalFile(), cp.persistTemp.getCanonicalFile());
+    assertEquals(expectedTarget.getCanonicalFile(), cp.getPersistTarget().getCanonicalFile());
+    assertEquals(expectedTemp.getCanonicalFile(), cp.getPersistTemp().getCanonicalFile());
 
-    // After initialization, the option's callback should report the current target path
     node.finishedInitialization();
     String configured = node.getString(optionName);
     assertEquals(expectedTarget.toString(), configured);
@@ -73,11 +73,10 @@ class ConfigurablePersisterTest {
     String optionName = "throttleFile";
     File baseDir = tmpDir.toFile();
 
-    // Create a file which we will try to treat as a parent directory
     File notADir = new File(baseDir, "parentFile");
     assertTrue(notADir.createNewFile(), "setup: parentFile should be created");
 
-    String defaultName = "parentFile" + File.separator + "child.txt"; // parent is a file
+    String defaultName = "parentFile" + File.separator + "child.txt";
 
     // Act + Assert
     NodeInitException ex =
@@ -97,7 +96,6 @@ class ConfigurablePersisterTest {
     assertEquals(NodeInitException.EXIT_THROTTLE_FILE_ERROR, ex.exitCode);
     String msg = ex.getMessage();
     assertNotNull(msg);
-    // Should contain the localized reason and mention the .tmp path
     String expectedReason =
         NodeL10n.getBase().getString("ConfigurablePersister.doesNotExistCannotCreate");
     assertTrue(msg.contains(expectedReason), msg);
@@ -106,7 +104,7 @@ class ConfigurablePersisterTest {
 
   @Test
   void set_whenChangeToNewValidPath_updatesTargetsAndCreatesFiles() throws Exception {
-    // Arrange: construct with a valid default
+    // Arrange
     Config cfg = new Config();
     SubConfig node = cfg.createSubConfig("node");
     String optionName = "throttleFile";
@@ -124,16 +122,16 @@ class ConfigurablePersisterTest {
 
     File newTarget = new File(baseDir, "new-throttles.txt");
 
-    // Act: update through the registered option (invokes callback -> setThrottles)
+    // Act
     Option<?> opt = node.getOption(optionName);
     opt.setValue(newTarget.toString());
 
-    // Assert: fields updated and files created
+    // Assert
     File expectedTemp = new File(newTarget + ".tmp");
     assertTrue(newTarget.exists(), "new target should be created");
     assertTrue(expectedTemp.exists(), "new temp should be created");
-    assertEquals(newTarget.getCanonicalFile(), cp.persistTarget.getCanonicalFile());
-    assertEquals(expectedTemp.getCanonicalFile(), cp.persistTemp.getCanonicalFile());
+    assertEquals(newTarget.getCanonicalFile(), cp.getPersistTarget().getCanonicalFile());
+    assertEquals(expectedTemp.getCanonicalFile(), cp.getPersistTemp().getCanonicalFile());
 
     node.finishedInitialization();
     assertEquals(newTarget.toString(), node.getString(optionName));
@@ -142,7 +140,7 @@ class ConfigurablePersisterTest {
   @Test
   void set_whenChangeToInvalidPath_throwsInvalidConfigValueAndDoesNotChangeTargets()
       throws Exception {
-    // Arrange: construct with a valid default
+    // Arrange
     Config cfg = new Config();
     SubConfig node = cfg.createSubConfig("node");
     String optionName = "throttleFile";
@@ -158,17 +156,16 @@ class ConfigurablePersisterTest {
                 baseDir),
             ticker);
 
-    File originalTarget = cp.persistTarget;
+    File originalTarget = cp.getPersistTarget();
 
-    // Create a file and attempt to use it as a parent directory for the new value
     File parentIsFile = new File(baseDir, "not-a-dir");
     assertTrue(parentIsFile.createNewFile(), "setup: not-a-dir should be created");
     String badPath = new File(parentIsFile, "child.txt").toString();
 
-    // Act + Assert: setting to an invalid path throws and does not update internal fields
+    // Act + Assert
     Option<?> opt = node.getOption(optionName);
     assertThrows(InvalidConfigValueException.class, () -> opt.setValue(badPath));
-    assertEquals(originalTarget.getCanonicalFile(), cp.persistTarget.getCanonicalFile());
+    assertEquals(originalTarget.getCanonicalFile(), cp.getPersistTarget().getCanonicalFile());
 
     node.finishedInitialization();
     assertEquals(originalTarget.toString(), node.getString(optionName));
