+++ src/test/java/network/crypta/runtime/persistence/PersisterTest.java
@@ -1,4 +1,4 @@
-package network.crypta.node;
+package network.crypta.runtime.persistence;
 
 import java.io.File;
 import java.io.FileOutputStream;
@@ -51,7 +51,6 @@ class PersisterTest {
     sfs.putSingle("nested.value", "ok");
     when(persistable.persistThrottlesToFieldSet()).thenReturn(sfs);
 
-    // No-op scheduler
     doAnswer(inv -> null)
         .when(ticker)
         .queueTimedJob(org.mockito.Mockito.any(Runnable.class), org.mockito.Mockito.anyLong());
@@ -61,13 +60,12 @@ class PersisterTest {
     // Act
     persister.run();
 
-    // Assert: file was written via temp then moved to target and contains our data
+    // Assert
     assertTrue(targetFile.isFile(), "persist target should exist");
     SimpleFieldSet readBack = SimpleFieldSet.readFrom(targetFile, false, true);
     assertEquals("42", readBack.get("alpha"));
     assertEquals("ok", readBack.get("nested.value"));
 
-    // Assert: ticker received a rescheduling for the same persister with the configured period
     ArgumentCaptor<Runnable> rCap = ArgumentCaptor.forClass(Runnable.class);
     ArgumentCaptor<Long> lCap = ArgumentCaptor.forClass(Long.class);
     verify(ticker, times(1)).queueTimedJob(rCap.capture(), lCap.capture());
@@ -77,7 +75,7 @@ class PersisterTest {
 
   @Test
   void run_whenParentDirectoryMissing_createsDirAndPersists() throws Exception {
-    // Arrange: point persister to a nested directory that doesn't exist yet
+    // Arrange
     File missingDir = new File(tempDir, "nested/missing");
     File tmp = new File(missingDir, "throttle.tmp");
     File tgt = new File(missingDir, "throttle.dat");
@@ -95,7 +93,7 @@ class PersisterTest {
     // Act
     persister.run();
 
-    // Assert: parent directory created, target written with expected content
+    // Assert
     assertTrue(tgt.isFile(), "persist target should exist after creating missing parent dirs");
     SimpleFieldSet readBack = SimpleFieldSet.readFrom(tgt, false, true);
     assertEquals("v", readBack.get("k"));
@@ -103,8 +101,7 @@ class PersisterTest {
 
   @Test
   void run_whenTempIsDirectory_logsAndSkipsWriteAndQueuesNext() {
-    // Arrange: make temp path a directory so FileOutputStream throws FileNotFoundException
-    // ("Is a directory")
+    // Arrange
     assertTrue(tempFile.mkdir(), "should create temp dir");
     SimpleFieldSet sfs = new SimpleFieldSet(false);
     sfs.putSingle("x", "y");
@@ -115,7 +112,7 @@ class PersisterTest {
     // Act
     persister.run();
 
-    // Assert: no valid file created at target (a directory rename might occur on some platforms)
+    // Assert
     assertFalse(
         targetFile.isFile(), "target should not be a regular file when temp is a directory");
     verify(ticker, times(1)).queueTimedJob(persister, Persister.PERIOD);
@@ -123,7 +120,7 @@ class PersisterTest {
 
   @Test
   void run_whenWriteThrows_deletesTempAndDoesNotMove() throws Exception {
-    // Arrange: return a SimpleFieldSet that throws during write
+    // Arrange
     SimpleFieldSet throwing = mock(SimpleFieldSet.class);
     doThrow(new IOException("boom")).when(throwing).writeToBigBuffer(org.mockito.Mockito.any());
     when(persistable.persistThrottlesToFieldSet()).thenReturn(throwing);
@@ -133,7 +130,7 @@ class PersisterTest {
     // Act
     persister.run();
 
-    // Assert: temp deleted in the IOException path, target not created
+    // Assert
     assertFalse(tempFile.exists(), "temp should be deleted after write failure");
     assertFalse(targetFile.exists(), "target should not be created when write fails");
     verify(ticker, times(1)).queueTimedJob(persister, Persister.PERIOD);
@@ -141,7 +138,7 @@ class PersisterTest {
 
   @Test
   void read_whenTargetExists_returnsParsedFieldSet() throws Exception {
-    // Arrange: write valid SFS to target directly
+    // Arrange
     SimpleFieldSet sfs = new SimpleFieldSet(false);
     sfs.putSingle("a", "b");
     try (FileOutputStream fos = new FileOutputStream(targetFile)) {
@@ -159,7 +156,7 @@ class PersisterTest {
 
   @Test
   void read_whenTargetUnreadableTempValid_returnsFromTemp() throws Exception {
-    // Arrange: create an empty target file so parsing throws EOFException and triggers fallback
+    // Arrange
     assertTrue(targetFile.createNewFile(), "should create empty target file");
     SimpleFieldSet sfs = new SimpleFieldSet(false);
     sfs.putSingle("k", "v");
@@ -171,7 +168,7 @@ class PersisterTest {
     // Act
     SimpleFieldSet read = persister.read();
 
-    // Assert: fallback to temp
+    // Assert
     assertNotNull(read);
     assertEquals("v", read.get("k"));
   }
@@ -190,7 +187,7 @@ class PersisterTest {
 
   @Test
   void read_whenBothUnreadable_returnsNull() throws Exception {
-    // Arrange: create empty files so both reads throw EOFException and return null
+    // Arrange
     assertTrue(targetFile.createNewFile(), "should create empty target file");
     assertTrue(tempFile.createNewFile(), "should create empty temp file");
     Persister persister = new Persister(persistable, tempFile, targetFile, ticker);
@@ -204,19 +201,18 @@ class PersisterTest {
 
   @Test
   void run_whenFatalThrowableThrown_catchesAndQueuesNext() {
-    // Arrange: simulate a fatal Error thrown during persistence
+    // Arrange
     doThrow(new LinkageError("simulated fatal error"))
         .when(persistable)
         .persistThrottlesToFieldSet();
 
     Persister persister = new Persister(persistable, tempFile, targetFile, ticker);
 
-    // Act: should not propagate the Error; still schedules next run
+    // Act
     persister.run();
 
-    // Assert: rescheduled even after fatal throwable
+    // Assert
     verify(ticker, times(1)).queueTimedJob(persister, Persister.PERIOD);
-    // And no target file created since write never happened
     assertFalse(targetFile.exists());
   }
 
@@ -228,11 +224,11 @@ class PersisterTest {
     when(persistable.persistThrottlesToFieldSet()).thenReturn(sfs);
     Persister persister = new Persister(persistable, tempFile, targetFile, ticker);
 
-    // Act: first start performs persist and schedules; second start is a no-op
+    // Act
     persister.start();
     persister.start();
 
-    // Assert: one schedule only; file exists with expected data
+    // Assert
     verify(ticker, times(1)).queueTimedJob(persister, Persister.PERIOD);
     assertTrue(targetFile.exists(), "persist should have executed once");
     SimpleFieldSet readBack = SimpleFieldSet.readFrom(targetFile, false, true);
